Update layout for Routing
Updated the layout configuration for the routing page.
The Map will be placed on the left side, and the locations
will be placed to the top-right of the page. Another box was
alsp placed in case we want to add any of the previously dicussed
features to the page. However, this is TBD.